fix: fix Slider (remove useState, add rem instead of px, refactor stories)
