+++ src/components/ui/slider/slider.module.scss
@@ -1,7 +1,7 @@
 .sliderWrapper {
   display: flex;
-  width: 251px;
-  height: 36px;
+  width: 15.688rem;
+  height: 2.25rem;
   justify-content: space-between;
   align-items: center;
 
@@ -9,8 +9,8 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    width: 36px;
-    height: 36px;
+    width: 2.25rem;
+    height: 2.25rem;
     border-radius: 2px;
     border: 1px solid var(--color-dark-300);
 
@@ -26,8 +26,8 @@
     align-items: center;
     user-select: none;
     touch-action: none;
-    width: 155px;
-    height: 16px;
+    width: 9.688rem;
+    height: 1rem;
 
     .track {
       background-color: var(--color-accent-900);
@@ -47,16 +47,21 @@
     .thumb {
       display: block;
       position: relative;
-      width: 16px;
-      height: 16px;
+      width: 1rem;
+      height: 1rem;
       background-color: var(--color-accent-500);
       border-radius: 10px;
+      cursor: pointer;
+
+      &:focus-visible {
+        outline: none;
+      }
 
       &::after {
         position: absolute;
         content: '';
-        width: 8px;
-        height: 8px;
+        width: 0.5rem;
+        height: 0.5rem;
         border-radius: 50%;
         top: 4px;
         left: 4px;
